feat: allow to retry temporary storage failures
Use a `RetryLayer` to make OpenDAL retry operations that aren't
permanent errors.  This can for example happen with S3 interactions
where the API returns a temporary error but the operation would succeed
if retried.  Doing the retry as part of the storage layer reduces the
chances of temporary failures to cause user-visible errors.

The configuration allows to use the retry-mechanism on file-system level
storage as well, though likely less applicable there.

Note that only temporary errors are retried, following the OpenDAL
classification of error classes, so operations that just will not
succeed are not retried needlessly and an error is returned immediately.